Disable ModelState validation in OnPostAsync method
The validation logic in the `OnPostAsync` method of the `Create.cshtml.cs` file within the `Apex.Events.Pages.GuestBookings`  namespace has been commented out.

Previously, the code checked if
the `ModelState` was valid, and if not, it populated the `GuestList` and `EventList` dropdowns and returned the page.

With this change, the `ModelState` validation and dropdown population logic are disabled, allowing the `GuestBooking` object to be added to the database context and saved regardless of  validation status.